Issue Fix #89 Appveyor: patching of Mkvcbuild.pm and vcregress.pl, should happen automatically
Currently, the msvc.diff file works on PostgreSQL 13 and master.

One is limited, in such a way, that one can not use msvc and the msvc.diff to
compile from source, PostgreSQL 12 (that is currently, still supported).
So, if someone should ask for a msvc "plr that is compiled from PostgreSQL 12 source",
one can not build that plr.

msvc.diff is used to modify the file Mkvcbuild.pm and vcregress.pl.
This file, Mkvcbuild.pm, is under heavy maintenance, and is modified using
the frequency, of once or twice per month.

https://github.com/postgres/postgres/commits/master/src/tools/msvc/Mkvcbuild.pm

In the master branch, in Mkvcbuild.pm (or vcregress.pl),
the situation is that every time someone (1) vertically shifts a line
or (2) or modifies of the Perl array @contrib_excludes,
one has to create a new msvc.diff file.

This fix enhances:

  1. avoids the manual re-creation of the msvc.diff file
     that would occur in Mkvcbuild.pm (and vcregress.pl) after
     (1) a vertical line shift or
     (2) addition or modification of the Perl array @contrib_excludes.
     by
     including a file msvc.diff.R that is a program that
     modifies the Mkvcbuild.pm and vcregress.pl.

  2. adds the ability to compile plr using PostgreSQL
     git branches, git tags, or git commit hashes

  3. adds demo compilations plr using PostgreSQL
     versions(git tags): REL_13_0 and REL_12_0

Issue Fix https://github.com/postgres-plr/plr/issues/89